pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Dmitriy Rogozov <kedovnet@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"

django-environ = "^0.8.1"
djangorestframework = "3.13.1"
djoser = "^2.1.0"
django-jazzmin = "^2.5.0"
whitenoise = "^6.2.0"
drf-spectacular = "^0.22.1"
psycopg2 = "^2.9.5"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
